requirements.txt
alembic==1.4.3
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.13.0
Flask-RESTful==0.3.8
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
pipenv==2020.6.2
pytest==5.4.3
python-dotenv==0.13.0
SQLAlchemy==1.3.17
tox==3.15.2
psycopg2-binary==2.8.5
bcrypt==3.2.0
